+++ chip-testing/src/scripts/Start.py
@@ -1,0 +1,3 @@
+#!/bin/sh
+echo "@ChipTestRunner: Command: 'Start' $@"
+sleep 3