+++ chip-testing/src/scripts/Stop.py
@@ -1,0 +1,3 @@
+#!/bin/sh
+echo "@ChipTestRunner: Command: 'Stop' $@"
+sleep 3