+++ chip-testing/src/suites-long/Test_TC_CADMIN_1_22.ts
@@ -5,13 +5,13 @@
  */
 
 // Needs to be first import!
-import { MinimalOnOffDeviceTestInstance } from "../MinimalOnOffDeviceTestInstance";
+import { Test_TC_CADMIN_1_4 } from "./Test_TC_CADMIN_1_4";
 
 import { StorageBackendMemory } from "@project-chip/matter.js/storage";
 
 /** Test case "Test_TC_CADMIN_1_22" */
-export class Test_TC_CADMIN_1_22 extends MinimalOnOffDeviceTestInstance {
+export class Test_TC_CADMIN_1_22 extends Test_TC_CADMIN_1_4 {
     constructor(storage: StorageBackendMemory, overrideTestName?: string) {
-        super(overrideTestName ?? "Test_TC_CADMIN_1_22", "GeneralTestPicsFile.txt", storage);
+        super(storage, overrideTestName ?? "Test_TC_CADMIN_1_22");
     }
 }