+++ chip-testing/src/suites-long/Test_TC_CADMIN_1_4.ts
@@ -14,4 +14,15 @@ export class Test_TC_CADMIN_1_4 extends MinimalOnOffDeviceTestInstance {
     constructor(storage: StorageBackendMemory, overrideTestName?: string) {
         super(overrideTestName ?? "Test_TC_CADMIN_1_4", "GeneralTestPicsFile.txt", storage);
     }
+
+    override async handleUserprompt(userPrompt: string, testDescription: string) {
+        if (testDescription.includes("Reset Devices to factory defaults")) {
+            return "y\n"; // Already done automatically
+        } else if (testDescription.includes("opens a commissioning window")) {
+            return "y\n";
+        } else if (testDescription.includes("Verify")) {
+            return "y\n";
+        }
+        return super.handleUserprompt(userPrompt, testDescription);
+    }
 }